% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_list_deliverability_test_reports}
\alias{pinpointemail_list_deliverability_test_reports}
\title{Show a list of the predictive inbox placement tests that you've
performed, regardless of their statuses}
\usage{
pinpointemail_list_deliverability_test_reports(NextToken, PageSize)
}
\arguments{
\item{NextToken}{A token returned from a previous call to \code{ListDeliverabilityTestReports}
to indicate the position in the list of predictive inbox placement
tests.}

\item{PageSize}{The number of results to show in a single call to
\code{ListDeliverabilityTestReports}. If the number of results is larger than
the number you specified in this parameter, then the response includes a
\code{NextToken} element, which you can use to obtain additional results.

The value you specify has to be at least 0, and can be no more than
1000.}
}
\description{
Show a list of the predictive inbox placement tests that you've
performed, regardless of their statuses. For predictive inbox placement
tests that are complete, you can use the \code{GetDeliverabilityTestReport}
operation to view the results.
}
\section{Request syntax}{
\preformatted{pinpointemail$list_deliverability_test_reports(
  NextToken = "string",
  PageSize = 123
)
}
}

\keyword{internal}
